% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{Your username}

\item{password}{Your password (minimum 12 characters)}
}
\value{

}
\description{
Authenticate
}
\examples{
\dontrun{
authenticate("username", "long-password-with-hyphens")
}
}
