% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_user}
\alias{create_user}
\title{Create a user}
\usage{
create_user(username, password)
}
\arguments{
\item{username}{The user's name}

\item{password}{The user's password}
}
\value{

}
\description{
Create a user
}
\examples{
\dontrun{
create_user("username", "the-password")
}
}
