% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{.apply_types}
\alias{.apply_types}
\title{Internal function - apply column types to dataframe}
\usage{
.apply_types(dataframe, columns)
}
\description{
Internal function - apply column types to dataframe
}
\keyword{internal}
