% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{.common_to_r_class}
\alias{.common_to_r_class}
\title{Internal function - return R type from abstract type}
\usage{
.common_to_r_class(common_type)
}
\description{
Internal function - return R type from abstract type
}
\keyword{internal}
