% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.do_http}
\alias{.do_http}
\title{Internal function - issue HTTP request}
\usage{
.do_http(method, endpoint, body = NULL, encode = NULL, timeout = NULL)
}
\description{
Internal function - issue HTTP request
}
\keyword{internal}
