% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{.get_columns}
\alias{.get_columns}
\title{Internal function - get column types for a data frame}
\usage{
.get_columns(dataframe)
}
\description{
Internal function - get column types for a data frame
}
\keyword{internal}
