% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.get_env}
\alias{.get_env}
\title{Internal function - get environment variable}
\usage{
.get_env(variable, default = "")
}
\description{
Internal function - get environment variable
}
\examples{
\dontrun{
get_env("AVATAR_BASE_URL")
}
}
\keyword{internal}
