% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes.R
\name{.r_to_common_types}
\alias{.r_to_common_types}
\title{Internal function - return abstract type from R type}
\usage{
.r_to_common_types(r_class)
}
\description{
Internal function - return abstract type from R type
}
\keyword{internal}
