% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forgotten_password}
\alias{forgotten_password}
\title{Request a password reset}
\usage{
forgotten_password(email)
}
\arguments{
\item{email}{The user's email}
}
\value{
Nothing
}
\description{
Request a password reset
}
\examples{
\dontrun{
forgotten_password("my.email@octopize.io")
}
}
