% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a dataset given a download_url and its columns}
\usage{
get_dataset(download_url, columns, download_timeout = 100)
}
\arguments{
\item{download_url}{The url to download the dataset}

\item{columns}{The types to apply to the dataset}

\item{download_timeout}{The time to wait for the download in seconds}
}
\value{
dataset
}
\description{
Get a dataset given a download_url and its columns
}
\examples{
\dontrun{
result <- get_job_result(job_id, timeout = get_result_timeout)
columns <- result$avatars_dataset$columns
download_url <- result$avatars_dataset$download_url
avatars <- get_dataset(download_url, columns)
}
}
