% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_explained_variance}
\alias{get_explained_variance}
\title{Get explained variance of the original records.}
\usage{
get_explained_variance(job_id)
}
\arguments{
\item{job_id}{The job id for this avatarization}
}
\value{
named list of the explained variance and the ratio
}
\description{
Get explained variance of the original records.
}
\examples{
\dontrun{
get_explained_variance(job_id)
}
}
