% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_job_result}
\alias{get_job_result}
\title{Get job result given a job id (blocking until the job is done)}
\usage{
get_job_result(job_id, timeout = 10)
}
\arguments{
\item{job_id}{The avatarization job id}

\item{timeout}{The time to wait for a result in seconds}
}
\value{
A list containing the job response
\itemize{
  \item{metrics}{a list containg security metrics, with names 'local_cloaking' and 'hidden_rate'}
  \item{avatars_dataset}{a list containing the avatars dataset metadata}
}
}
\description{
It can be safely retried multiple times until the job is done.
}
\examples{
\dontrun{
get_job_result(job_id)$metrics$local_cloaking
}
}
