% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_projections}
\alias{get_projections}
\title{Get projections of the original and avatars in original space.}
\usage{
get_projections(job_id)
}
\arguments{
\item{job_id}{The job id for this avatarization}
}
\value{
named list of original and avatar projection dataframes
}
\description{
Get projections of the original and avatars in original space.
}
\examples{
\dontrun{
get_projections(job_id)
}
}
