% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{healthcheck}
\alias{healthcheck}
\title{Check that the API is up and running}
\usage{
healthcheck()
}
\value{
Nothing
}
\description{
Check that the API is up and running
}
\examples{
\dontrun{
healthcheck()
}
}
