% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reset_password}
\alias{reset_password}
\title{Reset the password of an account}
\usage{
reset_password(email, new_password, new_password_repeated, token)
}
\arguments{
\item{email}{The user's email}

\item{new_password}{The new password}

\item{new_password_repeated}{The new password repeated}

\item{token}{A token received on the email of the account}
}
\value{
Nothing
}
\description{
Reset the password of an account
}
\examples{
\dontrun{
reset_password(
  email = "my.email@octopize.io",
  new_password = "b43b233a5e602b5a52ecab67cdf8",
  new_password_repeated = "b43b233a5e602b5a52ecab67cdf8",
  token = "8c85cf5a-809f-4b27-ba10-2495d5cfda5a"
)
}
}
