% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_server}
\alias{set_server}
\title{Configure the HTTP endpoint}
\usage{
set_server(server)
}
\arguments{
\item{server}{avatars API host}
}
\value{
Nothing
}
\description{
Configure the HTTP endpoint
}
\examples{
\dontrun{
set_server("http://localhost")
}
}
