% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{start_job}
\alias{start_job}
\title{Start the avatarization job}
\usage{
start_job(dataset_id, parameters)
}
\arguments{
\item{dataset_id}{An identifier for the uploaded dataset}

\item{parameters}{A list with avatarization parameters (see main tutorial)}
}
\value{
A job description (in particular with an id)
}
\description{
Start the avatarization job
}
\examples{
\dontrun{
start_job(dataset_id, list(k = 20))
start_job(dataset_id, list(k = 20, imputation = list(k = 5)))
start_job(dataset_id, list(k = 20, seed = 42, ncp = 3, column_weights = list("var_a" = 0.2)))
}
}
