% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_avatars}
\alias{get_avatars}
\title{Get avatars as a dataframe given a job id}
\usage{
get_avatars(job_id, get_result_timeout = 10, download_timeout = 100)
}
\arguments{
\item{job_id}{The job id for this avatarization}

\item{get_result_timeout}{The time to wait for the job result in seconds}

\item{download_timeout}{The time to wait for the download in seconds}
}
\value{
Avatars dataframe
}
\description{
The order of the lines have been shuffled, which means
that the link between original and avatar individuals cannot
be made.
}
\details{
Any further processing on the server will however use
the same avatar dataset as the one returned here.
}
\examples{
\dontrun{
get_avatars(job_id)
}
}
