% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upload_dataset}
\alias{upload_dataset}
\title{Upload dataset to the avatarization API}
\usage{
upload_dataset(dataframe, name = "original.csv")
}
\arguments{
\item{dataframe}{The original data}

\item{name}{The name to give the dataset}
}
\value{
A dataset id
}
\description{
Upload dataset to the avatarization API
}
\examples{
\dontrun{
dataset_id <- upload_dataset(df)
}
}
